# public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Mapa de Emergencia y Rescate — API",
    "version": "1.0.0",
    "description": "API pública/admin del mapa de emergencia. Documentación generada automáticamente desde los bloques @swagger de cada route."
  },
...
  "components": {
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/sync/duplicates": {
      "post": {
        "tags": [
          "sync"
        ],
        "summary": "Encola el reporte de posibles duplicados (requiere x-admin-token)",
        "parameters": [
          {
            "in": "query",
            "name": "source",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Filtra el reporte por id de fuente externa."
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "type": "integer"
            },
            "description": "Máximo de grupos de duplicados a devolver."
          }
        ],
        "responses": {
          "202": {
            "description": "Reporte encolado. Lee el resultado con GET /api/sync/status?jobId=.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "queued": {
                      "type": "boolean"
                    },
                    "jobId": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "No autorizado (falta o es inválido x-admin-token).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "No se pudo encolar.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/sync/geocode": {
      "get": {
        "tags": [
          "sync"
        ],
        "summary": "Geocodifica un lote de ubicaciones sin coordenadas (cron, requiere auth Bearer CRON_SECRET o token admin)",
        "parameters": [
          {
            "in": "query",
            "name": "max",
            "required": false,
            "description": "Tope de ubicaciones únicas a geocodificar en la corrida.",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Geocode encolado. Estado vía GET /api/sync/status?jobId=.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "queued": {
                      "type": "boolean"
                    },
                    "jobId": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "No autorizado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "No se pudo encolar.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/sync/run": {
      "post": {
        "tags": [
          "sync"
        ],
        "summary": "Dispara manualmente la sincronización de fuentes externas (requiere x-admin-token)",
        "parameters": [
          {
            "in": "query",
            "name": "dryRun",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "1",
                "true"
              ]
            },
            "description": "Si es 1/true simula la corrida sin escribir."
          },
          {
            "in": "query",
            "name": "source",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Limita la corrida a una sola fuente por su id."
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1
            },
            "description": "Tope de registros por fuente."
          },
          {
            "in": "query",
            "name": "mode",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "chunk"
              ]
            },
            "description": "Si es chunk procesa por páginas usando el cursor en sync_state."
          },
          {
            "in": "query",
            "name": "pages",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1
            },
            "description": "Tope de páginas por corrida (solo con mode=chunk)."
          }
        ],
        "responses": {
          "202": {
            "description": "Corrida ENCOLADA. El sync ya no corre inline (audit M-2): se procesa en el worker. Consulta el estado con GET /api/sync/status?jobId=.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "queued": {
                      "type": "boolean"
                    },
                    "jobIds": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Fuente desconocida.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "No autorizado (falta o es inválido x-admin-token).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "No se pudo encolar (cola no disponible).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/sync/status": {
      "get": {
        "tags": [
          "sync"
        ],
        "summary": "Estado de un job de sync encolado (status-poll, requiere x-admin-token)",
        "parameters": [
          {
            "in": "query",
            "name": "jobId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "jobId devuelto por POST /api/sync/run (o el cron)."
          }
        ],
        "responses": {
          "200": {
            "description": "Estado + resultado del job.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "jobId": {
                      "type": "string"
                    },
                    "state": {
                      "type": "string",
                      "description": "waiting|active|completed|failed|delayed"
                    },
                    "progress": {},
                    "result": {},
                    "failedReason": {
                      "type": "string",
                      "nullable": true
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Falta jobId.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "No autorizado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Job no encontrado (expiró o id inválido).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    }
  }
}